lots of import fixes and spelling corrections. unified the types (not two different enums in rule and labels)

+++ src/blossom/lang/instruction/InstructionGroup.java
@@ -1,5 +1,9 @@
 package blossom.lang.instruction;
 
+import java.util.ArrayList;
+
+import blossom.lang.Graph;
+
 public class InstructionGroup extends Instruction {
 
     public final ArrayList<Instruction> instructions;
@@ -14,7 +18,7 @@ public class InstructionGroup extends Instruction {
     }
 
     public Graph apply(final Graph g) {
-        int choice = Math.floor(Math.random() * instructions.size());
+        int choice = (int)Math.floor(Math.random() * instructions.size());
         System.out.printf("Choosing Instruction %s.\n", instructions.get(choice));
         return instructions.get(choice).apply(g);
     }
